Fixed issues with WPML v2.6.3
This is kind of a tricky update that I'm a littl unsure about. WPML was
having some issues with WordPress 3.5 and they quickly did a second
release for it, which was v2.6.3.

But with this update, the icl_get_languages function no longer works at
all in the wp-admin. So, I adjusted everything in the admin when I
retrieve languages to use the $sitepress object directly.

This resulted in some other chagnes that needed to made in order to
accomodate this. I have a strange feeling that this is a bug with WPML
that they'll eventually work out. So, I need to monitor these changes
through next updates from WPML. We may end up being able to roll
everything back to how it was in TB WPML Bridge v1.1.0.